Allow 'fixed' map to expand
